+++ Wallet/Views/Main/ContentView.swift
@@ -1,0 +1,32 @@
+//
+//  ContentView.swift
+//  Wallet
+//
+//  Created by Kamil Książek on 07/08/2022.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    @EnvironmentObject var viewModel: AuthViewModel
+
+    var body: some View {
+        Group {
+            if viewModel.userSession == nil {
+                LoginView()
+            } else {
+                MainView()
+            }
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        Group {
+            ContentView()
+            ContentView()
+                .preferredColorScheme(.dark)
+        }
+    }
+}
